pyproject.toml
[project]
name = "hotel-booking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiobotocore>=2.24.2",
    "alembic>=1.16.5",
    "asyncclick>=8.2.2.2",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "dishka>=1.7.1",
    "email-validator>=2.3.0",
    "environs>=14.3.0",
    "fastapi[standard]>=0.116.2",
    "phonenumberslite>=9.0.14",
    "pyjwt>=2.10.1",
    "python-magic>=0.4.27",
    "sqlalchemy>=2.0.43",
    "types-aiobotocore-lite[essential]>=2.24.2",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "faker>=37.8.0",
    "pytest>=8.4.2",
]

# Ruff configuration

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [".venv", "__pycache__", ".git", ".mypy_cache", "docker"]

[tool.ruff.lint.isort]
known-first-party = ["src.app", "src.contexts", "src.system", "src.shared", "src.utils"]
combine-as-imports = true
force-sort-within-sections = true
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

# Tests

[tool.pytest.ini_options]
python_files = "*.py"
python_functions = ["test_*", "check_*", "*_test"]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q"

# Mypy

[tool.mypy]
mypy_path = ["src"]
